Refactor ml/lstm package into subdirectories
