--- ml/lstm/dockerfile_train
@@ -1,17 +1,0 @@
-FROM python:3.11.6
-
-WORKDIR /opt/ml/code
-
-COPY ml/lstm/requirements_train.txt requirements.txt
-
-RUN pip3 install -r requirements.txt
-
-COPY ml/lstm/entrypoint_helpers.py .
-
-COPY pkg/storage pkg/storage
-
-COPY pkg/data pkg/data
-
-COPY ml/lstm/entrypoint_train.py .
-
-ENV SAGEMAKER_PROGRAM entrypoint_train.py
