+++ ml/lstm/predict/dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.11.6
 
-COPY ml/lstm/requirements_predict.txt requirements.txt
+COPY ml/lstm/requirements.txt requirements.txt
 
 RUN pip3 install -r requirements.txt
 
@@ -8,9 +8,9 @@ COPY pkg/storage pkg/storage
 
 COPY pkg/data pkg/data
 
-COPY ml/lstm/entrypoint_predict.py .
+COPY ml/lstm/predict/entrypoint.py .
 
-COPY ml/lstm/entrypoint_helpers.py .
+COPY ml/lstm/helpers/helpers.py .
 
 # conditionally copy model artifact for local builds
 COPY lstm.kera[s] model/
@@ -20,6 +20,6 @@ COPY scalers.pk[l] model/
 
 EXPOSE 8080
 
-ENTRYPOINT ["python3", "entrypoint_predict.py"]
+ENTRYPOINT ["python3", "entrypoint.py"]
 
 CMD ["serve"]
