+++ ml/lstm/train/dockerfile
@@ -1,0 +1,17 @@
+FROM python:3.11.6
+
+WORKDIR /opt/ml/code
+
+COPY ml/lstm/train/requirements.txt requirements.txt
+
+RUN pip3 install -r requirements.txt
+
+COPY ml/lstm/helpers/helpers.py .
+
+COPY pkg/storage pkg/storage
+
+COPY pkg/data pkg/data
+
+COPY ml/lstm/train/entrypoint.py .
+
+ENV SAGEMAKER_PROGRAM entrypoint.py
